RAG_documents/requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==1.0.12
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.14
docx==0.2.4
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.9
filelock==3.15.4
fitz==0.0.1.dev2
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.33.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.5
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.3
isodate==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==30.1.0
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.3.24
langchain-core==0.3.63
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langsmith==0.3.43
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
oauthlib==3.2.2
ollama==0.5.1
onnxruntime==1.19.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pillow==10.4.0
posthog==3.5.0
propcache==0.2.1
protobuf==4.25.4
prov==2.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==3.0.1
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.4
pypdf==5.6.0
PyPDF2==3.0.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.25.1
rsa==4.9
scipy==1.14.1
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.40.0
sympy==1.13.2
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tomli==2.0.1
tqdm==4.66.5
traits==6.3.2
typer==0.12.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.18.0
zipp==3.20.0
zstandard==0.23.0
